refactor(bundler): move signing utils out of app.rs
