+++ tooling/bundler/Cargo.toml
@@ -35,18 +35,18 @@ walkdir = "2"
 handlebars = { version = "3.5" }
 zip = { version = "0.5" }
 tempfile = "3.2.0"
-regex = { version = "1" }
+regex = "1"
 
 [target."cfg(target_os = \"windows\")".dependencies]
-attohttpc = { version = "0.17.0" }
-regex = { version = "1" }
+attohttpc = "0.17"
+regex = "1"
 uuid = { version = "0.8", features = [ "v5" ] }
 bitness = "0.4"
 winreg = "0.8"
 
 [target."cfg(not(target_os = \"linux\"))".dependencies]
-sha2 = { version = "0.9" }
-hex = { version = "0.4" }
+sha2 = "0.9"
+hex = "0.4"
 
 [lib]
 name = "tauri_bundler"
