+++ tooling/bundler/src/bundle/macos/mod.rs
@@ -1,2 +1,3 @@
 pub mod app;
 pub mod dmg;
+pub mod sign;
